Add skin support for sawblades
